examples/simple-py/requirements.txt
nats-py>=2.7.0
protobuf>=5.26.1
googleapis-common-protos>=1.63.0
